engine: extract esm/tours/tour_document.js
